1.[FORMATTING]

+++ vulkan-1.3/src/SceneNodeBuilder.cpp
@@ -381,7 +381,7 @@ void SceneNodeBuilder::processMeshes(fastgltf::Asset &gltf) {
 
         fastgltf::iterateAccessor<std::uint32_t>(
             gltf, indexaccessor,
-            [&](std::uint32_t idx) { indices.push_back(idx + initial_vtx); });
+            [&](std::uint32_t idx) { indices.push_back(idx + static_cast<uint32_t>(initial_vtx)); });
       }
 
       // load vertex positions
